services/api_gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
asyncpg==0.29.0
redis==5.0.4
httpx==0.27.0
structlog==24.1.0
python-dotenv==1.0.1
pydantic>=2.0
